% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mode.R
\name{calculate_mode}
\alias{calculate_mode}
\title{Determine majority (mode) of categorical variables in data frame}
\usage{
calculate_mode(x)
}
\arguments{
\item{x}{vector of values}
}
\value{

}
\description{
credits to https://stackoverflow.com/questions/55758611/how-to-find-the-most-frequent-value-from-multiple-columns-in-a-dataframe
}
