% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nngeo.R
\name{check_nngeo}
\alias{check_nngeo}
\title{check and assign potential geographic country to lat/lon position}
\usage{
check_nngeo(.sf_pts, .poly, .maxdist = 1e+05, ...)
}
\arguments{
\item{.sf_pts}{simple feature sf of position points (lat/lon)}

\item{.poly}{sf polygon(s) to verify whether points are in or next (.maxdist)
from polygon (e.g. aircraft position over country or waters close by)}

\item{.maxdist}{maximum distance in meters, e.g. 100km = 100000}
}
\value{
country code derived from .poly added to input sf_pts object
}
\description{
check and assign potential geographic country to lat/lon position
}
