% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_fst_coerce_date_to_character.R
\name{for_fst_coerce_date_to_character}
\alias{for_fst_coerce_date_to_character}
\title{Coerce timestamps to write out to fst}
\usage{
for_fst_coerce_date_to_character(.ds)
}
\arguments{
\item{.ds}{Opensky-Network flight table with day, firstseen, and lastseen}
}
\value{
tibble with coerced timestamps
}
\description{
Utility function to ensure character string for fst (binary file format).
}
