% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_dataset_description.R
\name{meta_dataset_description}
\alias{meta_dataset_description}
\title{OSN dataset meta description}
\usage{
meta_dataset_description(.ds, .yr)
}
\arguments{
\item{.ds}{OSN dataset}

\item{.yr}{year of observation}
}
\value{

}
\description{
OSN dataset meta description
}
