% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_latlon_to_sf.R
\name{pts_latlon_to_sf}
\alias{pts_latlon_to_sf}
\title{Coerce lat/lon points to sf}
\usage{
pts_latlon_to_sf(.ds, .lat_var, .lon_var, .remove = TRUE)
}
\arguments{
\item{.ds}{tibble of lat/lon pointsc, must have id column!}

\item{.lat_var}{latitude variable}

\item{.lon_var}{longitude variable}
}
\value{
sf object with points coerced to geometry
}
\description{
Coerce lat/lon points to sf
}
\examples{
\donotrun{
pts_latlon_to_sf(my_pts_tibble, lat, lon)
}
}
