% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_temp.R
\name{save_tmp}
\alias{save_tmp}
\title{save temp to store temporary results}
\usage{
save_tmp(.ds, .prefix, .pth = pth)
}
\arguments{
\item{.ds}{processed tibble requiring an (additional) id colum}

\item{.prefix}{a tag to recall what data subset is processed, e.g. departure side (DEP)}

\item{.pth}{path to store the temp files while iterating}
}
\value{
no return as file will be written out to pth folder
the function is chatty (ui_info)
}
\description{
This utility function saves temporary data prep results as fst into
a given subfolder of path, i.e. tmp!
The function also coerces the sf object to a normal data frame by
recovering the lat/lon and dropping the geometry. This allows the tibble
to be saved as fst (binary format).
}
\examples{
\dontrun{} save_tmp(rq, .prefix = "DEP")
}
