% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nngeo_and_save.R
\name{calc_nngeo_and_save}
\alias{calc_nngeo_and_save}
\title{Nearest neighbour and temp save for purrr iteration}
\usage{
calc_nngeo_and_save(.sf_pts, .poly, .prefix, .pth = pth, ...)
}
\arguments{
\item{.sf_pts}{}

\item{.poly}{world polygons}

\item{.prefix}{any code for the tmp folder}

\item{.pth}{path to store temp files}

\item{...}{any other useful stuff}
}
\value{

}
\description{
Wrapper function to have the (i) nearest neighbour and (ii) save while
iterating over large dataset. This is to reduce the risk of loosing the
work, if things fail or are switched off :)
The OSN data set for 2019 is about 7 mio flights without id for lat/lon.
That runs for days ...
}
\examples{
\donotrun{} calc_nngeo_and_save(add what you need)
}
